Tic-Tac-Toe Game
The user(x) plays Tic-Tac-Toe with the computer(o) in the console on a 3 x 3 game board. To play, the user will enter the coordinate he or she desires in a (x,y) setup. The computer will randomly generate an answer on the game board. The game will continue until a player wins or ends in a tie.