Popular Baby Names
The user can search a baby name when prompted. If the name is in the "name-data" file, the name will be printed in the console along with how popular the name was by decades. The asterisks represent the popularity of the name; therefor, the more asterisks there are, the more popular the name. The number at the end shows the rank of the name at the end of that decade. This is a good example of utilizing a queue.